chat_id BigInteger fix

+++ database.py
@@ -1,7 +1,7 @@
 """
 Database models and operations
 """
-from sqlalchemy import create_engine, Column, Integer, String, DateTime, Boolean, ForeignKey
+from sqlalchemy import create_engine, Column, Integer, BigInteger, String, DateTime, Boolean, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, relationship
 from datetime import datetime
@@ -20,7 +20,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer, unique=True, nullable=False, index=True)
+    chat_id = Column(BigInteger, unique=True, nullable=False, index=True)
     username = Column(String(255), nullable=True)
     is_active = Column(Boolean, default=True)
     is_subscribed = Column(Boolean, default=False)
